classifier-pipeline-resnet-quick-start/model-version-train-ps/1/models/model/requirements.txt
clarifai==12.1.4
clarifai-protocol>=0.0.23,<0.0.36
Pillow==12.0.0
PyYAML==6.0.3
mmcv==2.2.0
mmengine==0.10.7
mmpretrain==1.2.0
